chore: split mocks
